requirements.txt
langchain-anthropic
langchain
langchain_community
python-dotenv
streamlit
transformers
sentence_transformers
ipykernel
wikipedia
bs4
langchain-text-splitters
voyageai
langchain_voyageai
chromadb
langchain_huggingface
sentence_transformers
faiss-cpu
langsmith
langchain-groq
fastapi
langserve
sse_starlette
langchain_aws
validators
youtube_transcript_api
unstructured
wikipedia
pymupdf
pdfplumber
arxiv
duckduckgo-search
jq
langchain-nvidia-ai-endpoints